Rename data3 to data
